# src/main/java/com/example/empleados/controller/EmpleadoCreateController.java
package com.example.empleados.controller;

import com.example.empleados.dto.EmpleadoCreateRequest;
import com.example.empleados.dto.EmpleadoResponse;
import com.example.empleados.service.CreateEmpleadoService;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/empleados")
public class EmpleadoCreateController {

    private final CreateEmpleadoService createEmpleadoService;

    public EmpleadoCreateController(CreateEmpleadoService createEmpleadoService) {
        this.createEmpleadoService = createEmpleadoService;
    }

    @PostMapping
    public ResponseEntity<EmpleadoResponse> create(@Valid @RequestBody EmpleadoCreateRequest request) {
        EmpleadoResponse response = createEmpleadoService.create(request);
        return ResponseEntity.status(HttpStatus.CREATED).body(response);
    }
}